Create UDM_Arrays_Ex5
Exercise 5 Create  a new program to do  the following: a) declare an array numbers which will contain the following values:  5,  6, 17, 81, 43, 90, 66, 30, 76, 12     Decide yourself how you are going to initialise the array. 
 
b) Using a for loop write the code to     
- display all values in the array numbers     
- increment all elements in the array by 1 and display the values again     
- display last 5 values from the array numbers 
 
c) We can find the sum of all numbers in an array by:      
- declaring a variable to hold the sum    
 -  initialising the variable to zero    
 - iterating through the array and adding each number in turn to the existing sum 
 
The following code finds the sum of all integers in the array numbers: 
  int sum = 0;    
for ( i =0; i<numbers.length; i++)
{      
 sum = sum + numbers[i]; 
}
 
 
i) Add the code to calculate the sum of all elements in the array    numbers into your program and output the result. 
 
ii) Add more Java statements to your code to calculate the average     of the array numbers.Use the array attribute length in the      calculation. Output all results.      Why is the calculation of the average 'inaccurate'? 
 
e) re-write the code so that the program reads the array values from the keyboard.